StudentTeacherInfo.dart
PLP DART OOP2
This program defines two classes, Student and Teacher, each with their respective properties and a method to print out their information. The School class creates instances of Student and Teacher, and calls the printInfo() method for each. Finally, the main() function initializes an instance of the School class and calls the createPrintInfo() method.